requirements/app-all.txt
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
plotly==4.5.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.19
threadpoolctl==2.1.0
tqdm==4.49.0
tweepy==3.9.0
urllib3==1.25.10
vaderSentiment==3.3.2
Werkzeug==1.0.1
WTForms==2.3.3
